requirements.txt
fastapi
uvicorn
python-dotenv
selenium
langchain>=0.0.300
openai>=0.27.0
langchain-groq>=0.1.0
langchain-openai==1.0.1
httpx>=0.23.0